requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
docker-pycreds==0.4.0
Flask==2.0.3
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kaggle==1.5.12
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==6.1.0
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.5.6
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
termcolor==1.1.0
text-unidecode==1.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.10
Werkzeug==2.0.3
yaspin==2.1.0
